Add DESCRIPTION and MIT License
The DESCRIPTION file is the basic metadata that defines a directory as R package source.  I'll add more fields to it, but these are the basics.  I also added an MIT license.  I did these with `usethis::use_description()` and `usethis::use_mit_license("Matteo Vaiente")`.

Note that this adds the license in 2 forms.  The .md is preferred on GitHub, but not CRAN.  So LICENSE.md is included in .Rbuildignore, which excludes it and prevents a warning being thrown on package build.